third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGraphicsElement.getBBox-06-expected.txt

This is a testharness.js-based test.
[FAIL] Non-rendered symbol with rect returns zero bbox
  assert_equals: symbol1.getBBox().x expected 0 but got 50
[FAIL] display:none <g> with rect returns zero bbox
  assert_equals: g_none.getBBox().x expected 0 but got 70
Harness: the test ran to completion.
